refactor: refactor registry into service specific files
