+++ src/command/mod.rs
@@ -3,10 +3,7 @@ pub mod context;
 pub mod palette;
 pub mod registry;
 
-#[cfg(test)]
-mod tests;
-
 pub use commands::*;
 pub use context::*;
 pub use palette::*;
-pub use registry::*;
+pub use registry::CommandRegistry;
